+++ Sources/Moya/Plugin.swift
@@ -38,9 +38,12 @@ public protocol RequestType {
     /// Retrieve an `NSURLRequest` representation.
     var request: URLRequest? { get }
 
+    ///  Additional headers appended to the request when added to the session.
+    var sessionHeaders: [String: String] { get }
+
     /// Authenticates the request with a username and password.
-    func authenticate(user: String, password: String, persistence: URLCredential.Persistence) -> Self
+    func authenticate(username: String, password: String, persistence: URLCredential.Persistence) -> Self
 
     /// Authenticates the request with an `NSURLCredential` instance.
-    func authenticate(usingCredential credential: URLCredential) -> Self
+    func authenticate(with credential: URLCredential) -> Self
 }